+++ REvoDesign/tools/SessionMerger.py
@@ -1,9 +1,9 @@
 import os
 from pymol import cmd
 import argparse
-from REvoDesign.REvoDesign import logging as logger
+from REvoDesign import root_logger
 
-logging = logger.getChild(__name__)
+logging = root_logger.getChild(__name__)
 
 
 class PyMOLSessionMerger: