+++ REvoDesign/tools/client_tools.py
@@ -1,7 +1,7 @@
 import datetime
-from REvoDesign.REvoDesign import logging as logger
+from REvoDesign import root_logger
 
-logging = logger.getChild(__name__)
+logging = root_logger.getChild(__name__)
 
 import os
 import ssl