+++ CHANGELOG.md
@@ -6,6 +6,21 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- UI tips for installer
+- new Qt-free `ParallelExecutor`
+- `git` setup in installer, via `conda` or `winget`
+
+### Changed
+- global root logger and main logger
+- old `ParallelExecutor` -> `QtParallelExecutor`
+  
+### Fixed
+- working directory after window lauched
+
+### Removed
+- QtWidgets(`progressBar`) in Designer and Visualiser.
+
 
 ## [1.5.1] - 2024-02-28
 ### Added